back/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose[cryptography]
psycopg2-binary
python-multipart
python-dotenv
requests
passlib
argon2-cffi
google-cloud-aiplatform>=1.38.0
google-auth
pydantic-settings
